add hook useCreatePokeSkillSuggestion
